fix comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’

+++ hash_tables/6-hash_table_delete.c
@@ -8,7 +8,7 @@
 void hash_table_delete(hash_table_t *ht)
 {
 	hash_node_t *hash = NULL, *aux = NULL;
-	int i = 0;
+	long unsigned i = 0;
 
 	if (!ht)
 		return;
